MSI-2557: Investigate code duplication can be avoided.
