+++ InventoryInStorePickupQuoteGraphQl/etc/schema.graphqls
@@ -1,10 +1,10 @@
 # Copyright © Magento, Inc. All rights reserved.
 # See COPYING.txt for license details.
 
-input ShippingAddressInput {
+extend input ShippingAddressInput {
     pickup_location_code: String @doc(description: "The code of Pickup Location which will be used for In-Store Pickup.")
 }
 
-type ShippingCartAddress implements CartAddressInterface {
+extend type ShippingCartAddress {
     pickup_location_code: String @resolver(class: "\\Magento\\InventoryInStorePickupQuoteGraphQl\\Model\\Resolver\\ShippingAddress\\PickupLocationCode")
 }
